forecast: extract overlay rendering into forecast-overlay.js
history-graph.js was over the 600-line hard cap after the modeForecast
bars landed. Move drawForecastOverlay + drawForecastModeBars into a
new forecast-overlay.js module (pure side-effect — takes ctx + forecast
payload + window params, paints on the canvas). history-graph.js calls
it once with the resolved nowSec / cutoffSec; everything else stays
the same. No behaviour change.
